ci: init GitHub CI
The CI setup consists of a main CI and a nightly CI.

The night CI runs for every PR and push to the main branch
(open-source). It builds docker images, caches them, creates containers,
and runs the tests in the containers.

The nightly CI runs every night if the main branch has changed. It runs
the native build pipeline and tests.