refactor worker setup and rx/tx packet handling
Overall, we get rid of the worker_id and only use the lcore_id for identification.

We update the rx/tx packet handling as follows:
Each worker receives packets from all ports instead of just one to improve the distribution of the packet processing among all cores.
According to the port pairs defined in the parameters, packets are then forwarded to the correct port. The forwarding logic could be extended in the future.

Control plane traffic, such as ARP packets, are not anymore filtered using flow_api. This functionality should be replaced with mirror interfaces.

Furthermore, for simplicity, we removed the distributor process, which would handle the software based distribution of packets.