worker: add interface mirrors and refactor worker rx/tx handling (#13)
The new mirror library provides functionality to set up a mirror tap
interface for each physical interface and handle the forwarding of
packets from the physical interface to the mirror and vice versa. This
can be used, e.g., to forward ARP packets that are then handled by
Linux.
Fixes #7

We update the rx/tx packet handling as follows:
Each worker receives packets from all ports instead of just one to
improve the distribution of the packet processing among all cores.
According to the port pairs defined in the parameters, packets are then
forwarded to the correct port. The forwarding logic can be extended in
the future.

Furthermore, for simplicity, we removed the distributor process, which
would handle the software-based distribution of packets. We plan on
adding this later again.

Additionally, we get rid of the worker_id and only use the lcore_id for
identification for simplification.